Fixed issues w/ multiple targets

+++ Hunt/Get-CimAppCompatCache.ps1
@@ -1,6 +1,6 @@
 #requires -version 3
 
-function Get-CimAppCompatCache {
+Function Get-CimAppCompatCache {
 <#
 .SYNOPSIS
     Get shimcache execution artefacts from a remote computer.
@@ -54,13 +54,20 @@ function Get-CimAppCompatCache {
         $Protocol = 'Dcom'
     )
 
-    BEGIN {
+    Begin {
+        # Optionally check host reachability
         if ($Ping -and -not $(Test-Connection -Count 1 -Quiet -ComputerName $ComputerName)) {
             Write-Verbose "[$ComputerName] Host is unreachable."
-            break
+            continue
         }
 
+        # Init variables
         $cimOption = New-CimSessionOption -Protocol $Protocol
+        [uint32] $HKLM = 2147483650
+    }
+
+    Process {
+        # Init remote session
         try {
             if (-not $PSBoundParameters['ComputerName']) {
                 $cimSession = New-CimSession -SessionOption $cimOption -ErrorAction Stop -Verbose:$false
@@ -74,28 +81,25 @@ function Get-CimAppCompatCache {
         }
         catch [System.Management.Automation.PSArgumentOutOfRangeException] {
             Write-Warning "Alternative authentication method and/or protocol should be used with implicit credentials."
-            break
+            return
         }
         catch [Microsoft.Management.Infrastructure.CimException] {
             if ($Error[0].FullyQualifiedErrorId -eq 'HRESULT 0x8033810c,Microsoft.Management.Infrastructure.CimCmdlets.NewCimSessionCommand') {
                 Write-Warning "Alternative authentication method and/or protocol should be used with implicit credentials."
-                break
+                return
             }
             if ($Error[0].FullyQualifiedErrorId -eq 'HRESULT 0x80070005,Microsoft.Management.Infrastructure.CimCmdlets.NewCimSessionCommand') {
                 Write-Verbose "[$ComputerName] Access denied."
-                break
+                return
             }
             else {
                 Write-Verbose "[$ComputerName] Failed to establish CIM session."
-                break
+                return
             }
         }
-    }
 
-    PROCESS {
-        [uint32]$HKLM = 2147483650
+        # Process artefact collection
         $OS = Get-CimInstance -ClassName Win32_OperatingSystem -CimSession $cimSession -Verbose:$false
-
         if ($OS.Version -like "5.1*") { 
             $subKey = 'SYSTEM\CurrentControlSet\Control\Session Manager\AppCompatibility'
             $valueName = ''
@@ -104,7 +108,6 @@ function Get-CimAppCompatCache {
             $subKey = 'SYSTEM\CurrentControlSet\Control\Session Manager\AppCompatCache'
             $valueName = 'AppCompatCache'
         }
-
         if ($valueContent = (Invoke-CimMethod -Class 'StdRegProv' -Name 'GetBinaryValue' -Arguments @{hDefKey=$HKLM; sSubKeyName=$subKey; sValueName=$valueName} -CimSession $cimSession -Verbose:$false).uValue) {
             ConvertFrom-AppCompatCacheValue -CacheValue $valueContent -OSVersion $OS.Version -OSArchitecture $OS.OSArchitecture | ForEach-Object {
                 Write-Output ([pscustomobject] @{
@@ -116,8 +119,11 @@ function Get-CimAppCompatCache {
         }
     }
 
-    END {
-        Remove-CimSession -CimSession $cimSession
+    End {
+        # End remote session
+        if ($cimSession) {
+            Remove-CimSession -CimSession $cimSession
+        }
     }
 }
