+++ Hunt/Get-CimAsepLogon.ps1
@@ -1,6 +1,6 @@
 #requires -version 3
 
-function Get-CimAsepLogon {
+Function Get-CimAsepLogon {
 <#
 .SYNOPSIS
     Get logon artifacts on a remote computer.
@@ -53,13 +53,21 @@ function Get-CimAsepLogon {
         $Protocol = 'Dcom'
     )
 
-    BEGIN {
+    Begin {
+        # Optionally check host reachability
         if ($Ping -and -not $(Test-Connection -Count 1 -Quiet -ComputerName $ComputerName)) {
             Write-Verbose "[$ComputerName] Host is unreachable."
-            break
+            continue
         }
 
+        # Init variables
         $cimOption = New-CimSessionOption -Protocol $Protocol
+        [uint32] $HKLM = 2147483650
+        [uint32] $HKU = 2147483651
+    }
+
+    Process {
+        # Init remote session
         try {
             if (-not $PSBoundParameters['ComputerName']) {
                 $cimSession = New-CimSession -SessionOption $cimOption -ErrorAction Stop -Verbose:$false
@@ -73,29 +81,26 @@ function Get-CimAsepLogon {
         }
         catch [System.Management.Automation.PSArgumentOutOfRangeException] {
             Write-Warning "Alternative authentication method and/or protocol should be used with implicit credentials."
-            break
+            return
         }
         catch [Microsoft.Management.Infrastructure.CimException] {
             if ($Error[0].FullyQualifiedErrorId -eq 'HRESULT 0x8033810c,Microsoft.Management.Infrastructure.CimCmdlets.NewCimSessionCommand') {
                 Write-Warning "Alternative authentication method and/or protocol should be used with implicit credentials."
-                break
+                return
             }
             if ($Error[0].FullyQualifiedErrorId -eq 'HRESULT 0x80070005,Microsoft.Management.Infrastructure.CimCmdlets.NewCimSessionCommand') {
                 Write-Verbose "[$ComputerName] Access denied."
-                break
+                return
             }
             else {
                 Write-Verbose "[$ComputerName] Failed to establish CIM session."
-                break
+                return
             }
         }
 
-        [uint32]$HKLM = 2147483650
-        [uint32]$HKU = 2147483651
+        # Process artefact collection
         $SIDS = Invoke-CimMethod -Class 'StdRegProv' -Name 'EnumKey' -Arguments @{hDefKey=$HKU; sSubKeyName=''} -CimSession $cimSession -Verbose:$false | Select-Object -ExpandProperty sNames | Where-Object {$_ -match 'S-1-5-21-[\d\-]+$'}
-    }
 
-    PROCESS {
         $key = 'SYSTEM\CurrentControlSet\Control\Terminal Server\Wds\rdpwd'
         $value = 'StartupPrograms'
         if ($data = (Invoke-CimMethod -Class 'StdRegProv' -Name 'GetExpandedStringValue' -Arguments @{hDefKey=$HKLM; sSubKeyName=$key; sValueName=$value} -CimSession $cimSession -Verbose:$false).sValue) {
@@ -296,13 +301,16 @@ function Get-CimAsepLogon {
         }
     }
 
-    END {
-        Remove-CimSession -CimSession $cimSession
+    End {
+        # End session
+        if ($cimSession) {
+            Remove-CimSession -CimSession $cimSession
+        }
     }
 }
 
 function Local:Get-RegistryKey {
-    param(
+    Param (
         [Parameter(ValueFromPipelineByPropertyName = $True)]
         [Microsoft.Management.Infrastructure.CimSession[]]
         $CimSession,
