+++ Hunt/Get-CimDNSCache.ps1
@@ -74,43 +74,15 @@ function Get-CimDNSCache {
         $RecordType
     )
 
-    BEGIN {
+    Begin {
+        # Optionally check host reachability
         if ($Ping -and -not $(Test-Connection -Count 1 -Quiet -ComputerName $ComputerName)) {
             Write-Verbose "[$ComputerName] Host is unreachable."
-            break
+            continue
         }
 
+        # Init variables
         $cimOption = New-CimSessionOption -Protocol $Protocol
-        try {
-            if (-not $PSBoundParameters['ComputerName']) {
-                $cimSession = New-CimSession -SessionOption $cimOption -ErrorAction Stop -Verbose:$false
-            }
-            elseif ($Credential.Username) {
-                $cimSession = New-CimSession -ComputerName $ComputerName -Credential $Credential -Authentication $Authentication -SessionOption $cimOption -ErrorAction Stop -Verbose:$false
-            }
-            else {
-                $cimSession = New-CimSession -ComputerName $ComputerName -Authentication $Authentication -SessionOption $cimOption -ErrorAction Stop -Verbose:$false
-            }
-        }
-        catch [System.Management.Automation.PSArgumentOutOfRangeException] {
-            Write-Warning "Alternative authentication method and/or protocol should be used with implicit credentials."
-            break
-        }
-        catch [Microsoft.Management.Infrastructure.CimException] {
-            if ($Error[0].FullyQualifiedErrorId -eq 'HRESULT 0x8033810c,Microsoft.Management.Infrastructure.CimCmdlets.NewCimSessionCommand') {
-                Write-Warning "Alternative authentication method and/or protocol should be used with implicit credentials."
-                break
-            }
-            if ($Error[0].FullyQualifiedErrorId -eq 'HRESULT 0x80070005,Microsoft.Management.Infrastructure.CimCmdlets.NewCimSessionCommand') {
-                Write-Verbose "[$ComputerName] Access denied."
-                break
-            }
-            else {
-                Write-Verbose "[$ComputerName] Failed to establish CIM session."
-                break
-            }
-        }
-
         $sections = @{
             '1' = 'Answer' 
             '2' = 'Authority' 
@@ -132,7 +104,6 @@ function Get-CimDNSCache {
             '28' = 'AAAA' 
             '33' = 'SRV'
         }
-
         $filters = New-Object System.Collections.ArrayList
         switch ($PSBoundParameters.Keys) {
             "RecordName" {
@@ -175,7 +146,39 @@ function Get-CimDNSCache {
         }
     }
 
-    PROCESS {
+    Process {
+        # Init remote sessions
+        try {
+            if (-not $PSBoundParameters['ComputerName']) {
+                $cimSession = New-CimSession -SessionOption $cimOption -ErrorAction Stop -Verbose:$false
+            }
+            elseif ($Credential.Username) {
+                $cimSession = New-CimSession -ComputerName $ComputerName -Credential $Credential -Authentication $Authentication -SessionOption $cimOption -ErrorAction Stop -Verbose:$false
+            }
+            else {
+                $cimSession = New-CimSession -ComputerName $ComputerName -Authentication $Authentication -SessionOption $cimOption -ErrorAction Stop -Verbose:$false
+            }
+        }
+        catch [System.Management.Automation.PSArgumentOutOfRangeException] {
+            Write-Warning "Alternative authentication method and/or protocol should be used with implicit credentials."
+            return
+        }
+        catch [Microsoft.Management.Infrastructure.CimException] {
+            if ($Error[0].FullyQualifiedErrorId -eq 'HRESULT 0x8033810c,Microsoft.Management.Infrastructure.CimCmdlets.NewCimSessionCommand') {
+                Write-Warning "Alternative authentication method and/or protocol should be used with implicit credentials."
+                return
+            }
+            if ($Error[0].FullyQualifiedErrorId -eq 'HRESULT 0x80070005,Microsoft.Management.Infrastructure.CimCmdlets.NewCimSessionCommand') {
+                Write-Verbose "[$ComputerName] Access denied."
+                return
+            }
+            else {
+                Write-Verbose "[$ComputerName] Failed to establish CIM session."
+                return
+            }
+        }
+
+        # Process artefact collection
         Get-CimInstance -Namespace 'root/StandardCimv2' -ClassName 'MSFT_DNSClientCache' -Filter $filter -CimSession $cimSession -Verbose:$false | ForEach-Object {
             $obj = New-Object -TypeName psobject
             $obj | Add-Member -MemberType NoteProperty -Name 'ComputerName' -Value $ComputerName
@@ -191,7 +194,10 @@ function Get-CimDNSCache {
         }
     }
 
-    END {
-        Remove-CimSession -CimSession $cimSession
+    End {
+        # End session
+        if ($cimSession) {
+            Remove-CimSession -CimSession $cimSession
+        }
     }
 }
