+++ Hunt/Get-CimLogonHistory.ps1
@@ -1,6 +1,6 @@
 #requires -version 3
 
-function Get-CimLogonHistory {
+Function Get-CimLogonHistory {
 <#
 .SYNOPSIS
     Get logon history from a remote computer.
@@ -60,13 +60,23 @@ function Get-CimLogonHistory {
         $Identity
     )
 
-    BEGIN {
+    Begin {
+        # Optionally check host reachability
         if ($Ping -and -not $(Test-Connection -Count 1 -Quiet -ComputerName $ComputerName)) {
             Write-Verbose "[$ComputerName] Host is unreachable."
-            break
+            continue
         }
 
+        # Init variables
         $cimOption = New-CimSessionOption -Protocol $Protocol
+        $filter = $null
+        if ($Identity) {
+            $filter = "Caption='$Identity'"
+        }
+    }
+
+    Process {
+        # Init remote session
         try {
             if (-not $PSBoundParameters['ComputerName']) {
                 $cimSession = New-CimSession -SessionOption $cimOption -ErrorAction Stop -Verbose:$false
@@ -80,30 +90,24 @@ function Get-CimLogonHistory {
         }
         catch [System.Management.Automation.PSArgumentOutOfRangeException] {
             Write-Warning "Alternative authentication method and/or protocol should be used with implicit credentials."
-            break
+            return
         }
         catch [Microsoft.Management.Infrastructure.CimException] {
             if ($Error[0].FullyQualifiedErrorId -eq 'HRESULT 0x8033810c,Microsoft.Management.Infrastructure.CimCmdlets.NewCimSessionCommand') {
                 Write-Warning "Alternative authentication method and/or protocol should be used with implicit credentials."
-                break
+                return
             }
             if ($Error[0].FullyQualifiedErrorId -eq 'HRESULT 0x80070005,Microsoft.Management.Infrastructure.CimCmdlets.NewCimSessionCommand') {
                 Write-Verbose "[$ComputerName] Access denied."
-                break
+                return
             }
             else {
                 Write-Verbose "[$ComputerName] Failed to establish CIM session."
-                break
+                return
             }
         }
 
-        $filter = $null
-        if ($Identity) {
-            $filter = "Caption='$Identity'"
-        }
-    }
-
-    PROCESS {
+        # Process artefact collection
         Get-CimInstance -ClassName Win32_NetworkLoginProfile -Filter $filter -CimSession $cimSession -Verbose:$false | ForEach-Object {
             if ($_.UserId) {
                 if ($_.Privileges -eq 1) {
@@ -128,7 +132,10 @@ function Get-CimLogonHistory {
         }
     }
 
-    END {
-        Remove-CimSession -CimSession $cimSession
+    End {
+        # End session
+        if ($cimSession) {
+            Remove-CimSession -CimSession $cimSession
+        }
     }
 }
